+++ app/assets/stylesheets/modules/_purchase.scss
@@ -1,4 +1,102 @@
 .wrapper__purchase{
   min-height: 100vh;
   background: #f5f5f5;
+  .purchase__main{
+    background-color: #fff;
+    width: 700px;
+    margin: 0 auto;
+    &__header{
+      font-size: 22px;
+      padding: 32px;
+      text-align: center;
+      border-bottom: 1px solid #f5f5f5;
+      font-weight: 600;
+    }
+    &__content{
+      padding: 0 16px 40px;
+      max-width: 343px;
+      margin: 0 auto;
+
+      .buy__image{
+        padding-top: 32px;
+        display: inline-block;
+        &__url{
+          width: 60px;
+          height: 80px;
+          float: left;
+          img{
+            width: 60px;
+            height: 80px;
+          }
+        }
+        .buy__image__information{
+          margin-left: 80px;
+          padding-top: 10px;
+        }
+      }
+      .buy__price{
+        border-bottom: 1px solid #f5f5f5;
+        padding: 32px 0;
+        font-size: 20px;
+        font-weight: 600;
+        height: 30px;
+        &__title{
+          float: left;
+        }
+        &__main{
+          float: right;
+        }
+      }
+      .payment__method{
+        padding: 32px 0 8px 0;
+        &__title{
+          font-weight: 600;
+          float: left;
+        }
+        &__change{
+          text-align: right;
+          a{
+            color: #0099e8;
+            text-decoration: none;
+          }
+        }
+      }
+      .payment__method__information{
+        padding-bottom: 20px;
+        border-bottom: 1px solid #f5f5f5;
+      }
+      .delivery{
+        padding-top: 32px;
+        &__title{
+          font-weight: 600;
+          float: left;
+        }
+        .delivery__method__change{
+          text-align: right;
+          a{
+            color: #0099e8;
+            text-decoration: none;
+          }
+        }
+        .delivery__main{
+          padding-bottom: 32px;
+          border-bottom: 1px solid #f5f5f5;
+        }
+      }
+      form{
+        padding-top: 32px;
+        button{
+          display: block;
+          text-align: center;
+          background: #ea352d;
+          border: 1px solid #ea352d;
+          color: #fff;
+          width: 100%;
+          line-height: 48px;
+          font-size: 14px;
+          text-decoration: none;
+        }
+      }
+    }
+  }
 }
