+++ app/assets/stylesheets/products/_products_index.scss
@@ -176,40 +176,6 @@
   }
 }
 
-
-
-          .icon1 {
-            height: 68px;
-            margin-right: 20px;
-            width: 68px;
-            vertical-align: middle;
-          }
-          .icon2 {
-            height: 40px;
-            margin-right: 8px;
-            width: 144px;
-            vertical-align: middle;
-            cursor: pointer;
-          }
-          .icon3 {
-            height: 40px;
-            width: 144px;
-            vertical-align: middle;
-            cursor: pointer;
-          }
-        }
-      }
-       &__left {
-        display: block;
-        .icon1 {
-          height: 260px;
-          width: 300px;
-        }
-      }
-    }
-  }
-
-
 .main-header__bottom__right__new__user{
   a{
     background-color: rgb(234, 53, 45);
