--- app/views/purchase/done.html.haml
@@ -1,4 +1,0 @@
-%h2 購入が完了しました！
-%p Apple MacBook Pro 13インチ
-%p ¥135,000(送料込み)
-= link_to "メルカリをはじめる",destroy_user_session_path, method: :delete,class: "signup__submit__btn"
