+++ config/routes.rb
@@ -10,7 +10,6 @@ Rails.application.routes.draw do
   resources :users, only: [:show]
   resources :products
   resources :mypage
-  root to: "mypage#index"
 
   root to: "products#index"
   resources :users, only: [:edit,:show]
@@ -18,6 +17,11 @@ Rails.application.routes.draw do
   resources :products do
     collection do 
       get 'pagenotfound'
+      get 'done'
+    end
+    member do
+      get 'purchase'
+      post 'pay'
     end
   end
 
@@ -40,11 +44,4 @@ Rails.application.routes.draw do
     end
   end
 
-  resources :purchase, only: [:index] do
-    collection do
-      get 'index', to: 'purchase#index'
-      post 'pay', to: 'purchase#pay'
-      get 'done', to: 'purchase#done'
-    end
-  end
 end
